Extract keyshare and aggregator features to crates
