+++ packages/ciphernode/Cargo.toml
@@ -56,6 +56,7 @@ num = "0.4.3"
 net = { path = "./net" }
 rand_chacha = "0.3.1"
 rand = "0.8.5"
+router = { path = "./router" }
 serde = { version = "1.0.208", features = ["derive"] }
 serde_json = { version = "1.0.133" }
 sled = "0.34.7"
