+++ packages/ciphernode/aggregator/Cargo.toml
@@ -13,5 +13,6 @@ async-trait = { workspace = true }
 enclave-core = { path = "../core" }
 fhe = { path = "../fhe" }
 sortition = { path = "../sortition" }
+router = { workspace = true }
 data = { path = "../data" }
 tracing = { workspace = true }
