+++ packages/ciphernode/aggregator/src/lib.rs
@@ -1,3 +1,4 @@
+mod feature;
 mod plaintext_aggregator;
 mod publickey_aggregator;
 mod repo;
@@ -9,4 +10,5 @@ pub use publickey_aggregator::{
     PublicKeyAggregator, PublicKeyAggregatorParams, PublicKeyAggregatorState,
 };
 
+pub use feature::*;
 pub use repo::*;
