+++ packages/ciphernode/enclave_node/src/ciphernode.rs
@@ -10,11 +10,13 @@ use evm::{
     helpers::ProviderConfig, CiphernodeRegistryReaderRepositoryFactory, CiphernodeRegistrySol,
     EnclaveSolReader, EnclaveSolReaderRepositoryFactory,
 };
+use keyshare::KeyshareFeature;
 use logger::SimpleLogger;
 use net::{NetRepositoryFactory, NetworkManager};
 use rand::SeedableRng;
 use rand_chacha::rand_core::OsRng;
-use router::{CiphernodeSelector, E3RequestRouter, FheFeature, KeyshareFeature};
+use router::{E3RequestRouter, FheFeature};
+use sortition::CiphernodeSelector;
 use sortition::Sortition;
 use sortition::SortitionRepositoryFactory;
 use std::sync::{Arc, Mutex};
