+++ packages/ciphernode/keyshare/src/lib.rs
@@ -1,4 +1,6 @@
+mod feature;
 mod keyshare;
 mod repo;
+pub use feature::*;
 pub use keyshare::*;
 pub use repo::*;
