+++ packages/ciphernode/router/Cargo.toml
@@ -9,8 +9,8 @@ enclave-core = { path = "../core" }
 sortition = { path = "../sortition" }
 fhe = { path = "../fhe" }
 data = { path = "../data" }
-keyshare = { path = "../keyshare" }
-aggregator = { path = "../aggregator" }
+# keyshare = { path = "../keyshare" }
+# aggregator = { path = "../aggregator" }
 evm = { path = "../evm" }
 anyhow = { workspace = true }
 serde = { workspace = true }
