+++ packages/ciphernode/router/src/lib.rs
@@ -1,15 +1,11 @@
-mod ciphernode_selector;
 mod committee_meta;
 mod context;
 mod e3_request_router;
 mod hooks;
-// mod repositories;
 mod repo;
 
-pub use ciphernode_selector::*;
 pub use committee_meta::*;
 pub use context::*;
 pub use e3_request_router::*;
 pub use hooks::*;
-// pub use repositories::*;
 pub use repo::*;
