+++ packages/ciphernode/sortition/src/ciphernode_selector.rs
@@ -1,8 +1,8 @@
+use crate::{GetHasNode, Sortition};
 /// CiphernodeSelector is an actor that determines if a ciphernode is part of a committee and if so
 /// forwards a CiphernodeSelected event to the event bus
 use actix::prelude::*;
 use enclave_core::{CiphernodeSelected, E3Requested, EnclaveEvent, EventBus, Shutdown, Subscribe};
-use sortition::{GetHasNode, Sortition};
 use tracing::info;
 
 pub struct CiphernodeSelector {
