+++ packages/ciphernode/sortition/src/lib.rs
@@ -2,11 +2,13 @@
 #![crate_type = "lib"]
 // #![warn(missing_docs, unused_imports)]
 
+mod ciphernode_selector;
 mod distance;
 mod index;
 mod repo;
 mod sortition;
 
+pub use ciphernode_selector::*;
 pub use distance::*;
 pub use index::*;
 pub use repo::*;
